+++ frontend/src/App.vue
@@ -19,29 +19,8 @@
         </v-dialog>
       </v-layout>
 
-      <v-list two-line flat class="pa-0">
-        <v-card>
-          <v-list-item-group multiple>
-            <div v-for="(item, index) in items" :key="item.id">
-              <v-list-item>
-                <v-list-item-action @click="toggleCompleted(item)">
-                  <v-checkbox v-model="item.completed" color="primary"></v-checkbox>
-                </v-list-item-action>
-                <v-list-item-content>
-                  <v-list-item-title
-                    v-bind:class="{'primary--text done': item.completed}">{{ item.title }}
-                  </v-list-item-title>
-                </v-list-item-content>
-                <v-btn text icon color="red lighten-3" style="display: none" class="clear-btn"
-                       @click="clear(item, index)">
-                  <v-icon>mdi-close</v-icon>
-                </v-btn>
-              </v-list-item>
-              <v-divider></v-divider>
-            </div>
-          </v-list-item-group>
-        </v-card>
-      </v-list>
+      <todo-list></todo-list>
+
     </v-container>
 
     <v-snackbar v-model="snackbar.show" :color="snackbar.color" :bottom="true" :right="true" :timeout=6000>
@@ -54,12 +33,14 @@
 <script>
   import axios from "./plugins/axiosbase";
   import TodoTextInput from "./components/todo-text-input"
+  import TodoList from "./components/todo-list"
   import {mapMutations, mapState} from "vuex";
 
   export default {
     name: 'App',
     components: {
-      TodoTextInput
+      TodoTextInput,
+      TodoList
     },
     data: function () {
       return {
@@ -88,16 +69,6 @@
           for (let i = 0; i < response.data.length; i++) {
             this.items.push(response.data[i])
           }
-        })
-      },
-      toggleCompleted: function (toggled) {
-        axios.put("/todo" + "/" + toggled.id, {completed: toggled.completed})
-      },
-      clear: function (item, index) {
-        axios.delete("/todo" + "/" + item.id).then(() => {
-          this.items.splice(index--, 1)
-        }).catch(() => {
-          this.showErrorSnackbar();
         })
       },
       clearCompleted: function () {
